requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
bs4==0.0.1
cmake==3.17.3
cycler==0.10.0
decorator==4.4.2
Django==3.0.7
dlib==19.20.0
glob3==0.0.1
image==1.5.32
imageio==2.8.0
kiwisolver==1.2.0
matplotlib==3.2.2
networkx==2.4
numpy==1.19.0
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tifffile==2020.6.3
